Fix ServletSymbol::yield
Aligning it better with the other implementations, so that it returns the
best option for AdaptationResult::argument when a value can be found in a
request, but can't be adapted and the context does not provide any valid
substitute. Then the argument does not contain the request's value, but
rather the context's value, although the request's value should be
preferred.